refactor: getFirstSeries
